1.0.2.3
Song Editor is now working in the tests that I've done.

Added a base class for all song editor controls to descend from.
Updated ucSongInfo and ucTones controls. Both are working.

Merged the AppSettings and SettingsData in to one class and inherited
AppSettings to a new class "NotifyPropChangedBase"
Added a SectionCount in the SongData.Arrangement class and added it to
the DetailsGrid in the SongManager control.
Added a "LogMessages" property to AppSettings, disabling speeds up song
parsing significantly.